// src/interfaces/Users/Users.ts
export interface UserProfile {
    id: number;
    path?: string;
    name?: string;
    role?: string;
    email?: string;
    dateOfBirth?: string;
    phone?: string;
    gender: string;
    subscription?: string;
    status: string;
    maritalStatus?: string;
    tribe?: string;
    employmentStatus?: string;
    reported?: boolean;
    blocked?: boolean
}